removing GareGeometry
extending Gare from MasonGeometry
